requirements.txt
allure-pytest==2.8.10
allure-python-commons==2.8.10
altgraph==0.16.1
apipkg==1.5
app==0.0.1
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.4
atomicwrites==1.3.0
attrs==19.3.0
aws-requests-auth==0.4.2
bcrypt==3.1.7
black==19.10b0
certifi==2019.3.9
cffi==1.13.2
chardet==3.0.4
Click==7.0
codegen==1.0
colorama==0.4.1
colorlog==4.1.0
cryptography==3.4.7
ddt==1.2.1
django-redis==4.12.1
docopt==0.6.2
enum34==1.1.10
et-xmlfile==1.0.1
execnet==1.7.1
Flask==1.1.1
Flask-WTF==0.14.3
future==0.18.2
gevent==1.4.0
greenlet==0.4.15
idna==2.8
importlib-metadata==1.5.0
iniconfig==1.0.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
jmespath==0.9.5
jsonpath==0.82
jwt==1.2.0
lml==0.0.9
logger==1.4
loguru==0.4.1
lxml==4.5.2
MarkupSafe==1.1.1
more-itertools==8.2.0
Naked==0.1.31
namedlist==1.7
nose==1.3.7
openpyxl==3.0.3
packaging==20.1
paho-mqtt==1.5.0
paramiko==2.7.2
passlib==1.7.4
pathspec==0.8.1
patterns==0.3
pefile==2019.4.18
ping3==2.6.1
pluggy==0.13.1
prettytable==0.7.2
py==1.9.0
pycparser==2.19
pycryptodome==3.10.1
pydantic==1.7.3
pyecharts==1.6.2
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-xlsx==0.5.8
pyftpdlib==1.5.6
PyInstaller==3.5
PyJWT==2.1.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.6
pytest==5.4.3
pytz==2019.3
pywin32-ctypes==0.2.0
PyYAML==5.3.1
records==0.5.3
redis==3.5.3
regex==2020.11.13
requests==2.22.0
selenium==3.141.0
sentry-sdk==0.14.4
shellescape==3.8.1
simplejson==3.17.0
six==1.13.0
SQLAlchemy==1.3.15
sqlparse==0.4.1
sshtunnel==0.4.0
tablib==1.1.0
texttable==1.6.2
toml==0.10.1
tornado==6.0.3
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.0.0
unrar==0.4
urllib3==1.24.1
wcwidth==0.1.8
websocket-client==0.56.0
Werkzeug==0.16.0
win32-setctime==1.0.1
WTForms==2.3.1
zipp==2.2.0
